requirements.txt
psutil~=6.1.0
python-telegram-bot~=21.7
APScheduler~=3.10.4
pytest~=8.3.3
requests~=2.32.3